ScalarHandlers separated
